requirements.txt
# for wd14-tagger
deepdanbooru
onnxruntime-silicon; sys_platform == 'darwin'
onnxruntime-gpu; sys_platform != 'darwin'
numba; sys_platform != 'darwin'
fastapi
gradio
huggingface_hub
numpy
opencv_contrib_python
opencv_python
opencv_python_headless
packaging
pandas
Pillow
tensorflow
tensorflow_io
tqdm

# for sd-webui-infinite-image-browsing #
fastapi
uvicorn
piexif
python-dotenv
Pillow